Restructure Nats module to support both connectors
