+++ stdlib/messaging/nats/src/main/ballerina/nats/basic/message.bal
