+++ stdlib/messaging/nats/src/main/java/org/ballerinalang/nats/Utils.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.nats.nativeimpl;
+package org.ballerinalang.nats;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.bre.bvm.BLangVMErrors;
