+++ stdlib/messaging/nats/src/main/java/org/ballerinalang/nats/basic/producer/RequestReply.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.nats.nativeimpl.producer;
+package org.ballerinalang.nats.basic.producer;
 
 import io.nats.client.Message;
 import io.nats.streaming.StreamingConnection;
@@ -31,8 +31,8 @@ import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.model.values.BValueArray;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.nats.nativeimpl.Constants;
-import org.ballerinalang.nats.nativeimpl.Utils;
+import org.ballerinalang.nats.Constants;
+import org.ballerinalang.nats.Utils;
 
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.CompletableFuture;
